cukinia-tests: add cukinia-all.conf
Currently, when calling cukinia alone, all the tests, including future
tests are run. As future tests are tests on security measures not
implemented yet, this will output tests failures.

To prevent this, all the tests are now called with a new file
`cukinia-all.conf`, and the tests run with Cukinia by default are those
set in `cukinia.conf` file.

Signed-off-by: Paul Le Guen de Kerneizon <paul.leguendekerneizon@savoirfairelinux.com>