common_security_tests.d/apt.conf: rework apt repository tests
The Ansible variable apt_repo has been converted into a list, which now
includes the URLs of all apt repositories. The revised tests for apt
repositories now verify all repositories specified in apt_repo.
Additionally, only the specified repositories are configured as apt
sources.

If apt_repo is an empty list, the tests will ensure that no repository
is configured as an apt source.

In cases where apt_repo is not defined, the apt repositories tests will
be skipped.

Signed-off-by: Mathieu Dupré <mathieu.dupre@savoirfairelinux.com>