workflows/certify-commit.yml: remove file
The old test process to devellop on cukinia-tests was :
- Review the Pull request on cukinia-tests
- Create a pull request on ansible that bump the cukinia-tests revision
- Pass the CI on ansible (this way, the PR on cukinia-tests was tested)
- Merge on ansible
- Merge on cukinia-tests

However, this implied that the hash of the cukinia-tests PR changed at
the last step (because of rebase). Therefore, the hash for the submodule
on the ansible repository was not right anymore.

The new workflow is now :
- Review and merge the PR on cukinia-tests
- Create a PR on ansible that bump the cukinia-tests revision
- If the CI fails, correct the behaviour by adding commit on
  cukinia-tests
- merge the PR on ansible

This workflow doesn't protect the main branch anymore and commit that
break the CI may be merged on it. But because the revision of
cukinia-tests on the ansible repository is a specific hash and not the
main branch, it shouldn't cause so many problems.

This commit remove the old workflow file.

Signed-off-by: Erwann Roussy <erwann.roussy@savoirfairelinux.com>