systemd.conf: fix services list bug
This commit fixes a bug in the list of essential services, where the string |\ had be wrongly typed \|
This make the whole "grep" operation fail, and the test was successful all the time...
Correcting this bug led to discover that a few essential services (polkit, chrony-wait...) had be either forgotten or entered with the ".service" extension, which is wrong.
This commit fixes all that.

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>