SEAPATH-00194: add overlay2 exception
Docker overlay files have the drwxrwxrwt permission which matches test 00194 (writable by world, but not belonging to root user).

The world writable permission is harmless because the parent directories are protected against unwanted access (the executable bit is turned off) and so the directories cannot be traversed/accessed to arrive to such world-writable directories or files.

https://www.ibm.com/support/pages/some-icp-filesdirectories-are-created-world-writable-permission

Signed-off-by: Florent CARLI <florent.carli@rte-france.com>